Create retraction_watch_by_country.R
Shiny dashboard for retraction watch data set. Parses data using custom categorization, cleans state field, and then constructs new columns and aggregates as part of dashboard ETL. The dashboard itself uses the aggregated table made by the ETL so it does not have to re-calc each change. The original data set can be found at the URL on line 1 (it has nearly 50k rows as of 11/27/23), and the original R script which inspired the dashboard can be found at the URL on line 2.